fix: move resolution utils into right folder
