+++ packages/test-apps/react-vite/src/binding-to-reference-fields/resolutionUtils/index.ts
