+++ packages/test-apps/react-vite/src/binding-to-reference-fields/resolutionUtils/resolution.ts
